fix: change morse_encode return types to size_t

+++ src/morse.c
@@ -48,12 +48,12 @@ const static chartbl_lookup_row morse_code[] = {
   {')', "-.--.- "},
 };
 
-int textproc_morse_encode(char* str, char* out, size_t outsiz)
+size_t textproc_morse_encode(char* str, char* out, size_t outsiz)
 {
   return textproc_morse_encode_n(str, out, outsiz, strlen(str));
 }
 
-int textproc_morse_encode_n(char* str, char* out, size_t outsiz, size_t bufsiz)
+size_t textproc_morse_encode_n(char* str, char* out, size_t outsiz, size_t bufsiz)
 {
   char lower_str[bufsiz];
   int rc;
